% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_uncertainty.R
\name{aggregate_uncertainty}
\alias{aggregate_uncertainty}
\title{Aggregate Bootstrap Uncertainty (Now includes Median Central Estimate)}
\usage{
aggregate_uncertainty(boot_dt, prob_col)
}
\arguments{
\item{boot_dt}{The combined data.table from all bootstrap iterations}

\item{prob_col}{The name of the probability column (e.g., "p_quit")}
}
\description{
Aggregate Bootstrap Uncertainty (Now includes Median Central Estimate)
}
