% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_bootstrap_sample.R
\name{generate_bootstrap_sample}
\alias{generate_bootstrap_sample}
\title{Generate a Single Bootstrap Sample for Complex Survey Data}
\usage{
generate_bootstrap_sample(survey_data)
}
\arguments{
\item{survey_data}{A data.table or data.frame containing survey data.}
}
\value{
A resampled data.table.
}
\description{
Generate a Single Bootstrap Sample for Complex Survey Data
}
