% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkins_relapse.R
\docType{data}
\name{hawkins_relapse}
\alias{hawkins_relapse}
\title{Long-term probabilities of smoking relapse}
\format{
A data table.
}
\source{
\insertRef{hawkins2010long}{smktrans}
}
\usage{
hawkins_relapse
}
\description{
Probabilities of smoking relapse estimated from Hawkins 2010, who analysed the British Household Panel Survey.
}
\keyword{datasets}
