% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd_data.R
\docType{data}
\name{hmd_data_eng}
\alias{hmd_data_eng}
\title{Death rates for England & Wales}
\format{
A data table.
}
\source{
\href{https://www.mortality.org/cgi-bin/hmd/country.php?cntr=GBRCENW&level=1}{Human Mortality Database - England and Wales}
}
\usage{
hmd_data_eng
}
\description{
Data from the Human Mortality Database on the 1x1 age and period central death rates (mx)
for the England & Wales civilian population. Death rates are stratified by sex only. Ages 11-89 selected
}
\keyword{datasets}
