% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd_data.R
\docType{data}
\name{hmd_data_scot}
\alias{hmd_data_scot}
\title{Death rates for Scotland}
\format{
A data table.
}
\source{
\href{https://www.mortality.org/cgi-bin/hmd/country.php?cntr=GBR_SCO&level=1}{Human Mortality Database - Scotland}
}
\usage{
hmd_data_scot
}
\description{
Data from the Human Mortality Database on the 1x1 age and period central death rates (mx)
for the Scotland population. Death rates are stratified by sex only. Ages 16-89 selected
}
\keyword{datasets}
