% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_est.R
\name{init_est}
\alias{init_est}
\title{Cohort specific smoking initiation \lifecycle{stable}}
\usage{
init_est(data, strat_vars = c("sex", "imd_quintile"))
}
\arguments{
\item{data}{Data table of individual characteristics.}

\item{strat_vars}{Character vector - the variables by which to 
stratify the cohort estimates e.g. sex and/or
IMD quintile.}
}
\value{
Returns a data table containing the age-specific probabilities and 
cumulative probabilities of
starting to smoke, and of being an ever or never smoker.
}
\description{
Calculate the age-specific probabilities and cumulative probabilities of 
starting to smoke cigarettes regularly
based on self-reported smoking histories.
}
\details{
Takes the self-reported age that cigarettes were first smoked regularly 
and constructs a vector of 0s, 1s and NAs
corresponding to each age for each individual. 1 = initiated smoking, NA = censored.
}
\examples{

\dontrun{

test_data <- init_est(data = hse_data)

}

}
