% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_relapse.R
\name{prep_relapse}
\alias{prep_relapse}
\title{Prepare long-term relapse probabilities \lifecycle{stable}}
\usage{
prep_relapse(
  data,
  hawkins_relapse = smktrans::hawkins_relapse,
  lowest_year = 2001,
  highest_year = 2018,
  youngest_age = 11
)
}
\arguments{
\item{data}{Data table containing individual characteristics from the Health Survey for England.}

\item{hawkins_relapse}{Data table containing a tidied version of the estimates of
long-term smoking relapse probability
 from \insertCite{hawkins2010long;textual}{smktrans}.}

\item{lowest_year}{integer - lowest year of data available (for England this is 2001
and for Scotland this is 2008). Default is set to 2001, for HSE.}

\item{highest_year}{integer - highest year of data available.}

\item{youngest_age}{integer - youngest age in data (for England we use 11, but for
Scotland this is 16).}
}
\value{
Returns two data tables: First, with relapse probabilities stratified by year,
age, IMD quintile and time since quit
(for use in the STAPM simulation);
 Second, with relapse probabilities stratified by just year, age and IMD
 quintile (for use in transition prob estimation).
}
\description{
Combines published estimates of long-term relapse with the Health Survey for England data
to arrive at the expected values for relapse probabilities within defined subgroups.
}
\details{
This function takes the estimates of relapse to smoking from 
\insertCite{hawkins2010long;textual}{smktrans}
and to process them into probabilities of relapse to smoking by the significant
variables from the above paper (age, time since quit, degree or not, mental health condition or not,
married or not). Note that physical health / gp visits was also significant but not included here
partly because the Health Survey for England doesn't have the right variables in all years
and partly because it might not be right to include this when we are looking at
health as an outcome in the model.
Once we have mapped relapse prob onto the hse by the above variables,
 we can then calculate the variation in expected probability of relapse by imd quintile.
}
\examples{
\dontrun{
test_data <- prep_relapse(
  data = hse_data,
  hawkins_relapse = smktrans::hawkins_relapse
)
}
}
\references{
\insertRef{hawkins2010long}{smktrans}
}
