% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_surv.R
\name{prep_surv}
\alias{prep_surv}
\title{Estimate socioeconomic differences in cohort survivorship}
\usage{
prep_surv(
  mx_data_hmd = smktrans::hmd_data,
  mx_data_ons = smktrans::tob_mort_data
)
}
\arguments{
\item{mx_data_hmd}{Data table containing mortality rates from 1922
from the Human Mortality Database.}

\item{mx_data_ons}{Data table containing socioeconomically stratified
mortality rates - that we process ourselves
from mortality microdata supplied by the Office for National Statistics.}
}
\value{
Returns a data table containing the socioeconomically stratified cohort survivorship functions.
Note that these data will only be stratified by IMD quintile for
year ages and years covered by our ONS data.
}
\description{
Takes mortality rates from birth since 1922 and combines
them with socioeconomically stratified mortality
data from 2001 to produce an estimate of the socioeconomic differences in survivorship.
}
\examples{
\dontrun{
test_data <- prep_surv(
  mx_data_hmd = smktrans::hmd_data,
  mx_data_ons = smktrans::tob_mort_data
)
}
}
