% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoke_surv.R
\name{smoke_surv}
\alias{smoke_surv}
\title{Estimate age-specific probabilities of death by smoking status \lifecycle{stable}}
\usage{
smoke_surv(
  data,
  diseases = tobalcepi::tob_disease_names,
  mx_data,
  min_age = 11,
  max_age = 89,
  min_year = 2003,
  max_year = 2018
)
}
\arguments{
\item{data}{Data table of individual characteristics, including age, sex, 
IMD quintile and smoking status.}

\item{diseases}{Character vector of smoking related conditions.}

\item{mx_data}{Data table containing the cause-specific rates of death 
from smoking related diseases.}

\item{min_age}{Integer - the youngest age considered.}

\item{max_age}{Integer - the oldest age considered.}

\item{min_year}{Integer - the earliest year of observed data.}

\item{max_year}{Integer - the latest year of observed data.}
}
\value{
Returns two data tables containing the age-specific 
probabilities of death in and survival through a
1 year age interval, stratified by sex, year, imd_quintile and smoking state. 
One data table has the detailed estimates and the
other has been tidied into the form needed for quit probability estimation.
}
\description{
Estimates the probility that an individual will survive 
from the beinning to the end of a year interval
with different proabilities of survival depending on an individuals smoking status
where the probability of survival for former smokers is adjusted according to our estimates
of how risk declines by the time since some has quit.
}
\examples{

\dontrun{

test_data <- smoke_surv(
  data = hse_data,
  diseases  = tobalcepi::tob_disease_names,
  mx_data = tob_mort_data_cause
)

}

}
