% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_var.R
\name{bin_var}
\alias{bin_var}
\title{Bin numeric variable}
\usage{
bin_var(x, n_bins)
}
\arguments{
\item{x}{Numeric variable to be binned}

\item{n_bins}{Integer - the number of bins to create}
}
\value{
Returns a numeric binned variable
}
\description{
Converts a numeric variable to bins of specified number, where the binned variable is also numeric and has the mid-points of the bins.
}
\examples{
bin_var(1:10, 3)


}
