% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_reports.R
\name{build_web_reports}
\alias{build_web_reports}
\title{Build Web Reports for Pkgdown}
\usage{
build_web_reports(config_list)
}
\arguments{
\item{config_list}{A list of configuration objects (e.g. list(config_eng, config_scot))}
}
\description{
Generates HTML articles for the package website based on the
configuration lists used in the analysis.
}
