% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_quitting.R
\name{estimate_quitting}
\alias{estimate_quitting}
\title{Estimate and Forecast Smoking Quitting}
\usage{
estimate_quitting(
  config,
  survey_data,
  tob_mort_data,
  tob_mort_data_cause,
  boot_mode = FALSE,
  smk_init_data_boot = NULL,
  relapse_data_boot = NULL,
  precalc_mortality = NULL
)
}
\description{
1. Trend fitting & Mortality calculation.
2. Historical Quit Solver.
3. Forecasts Quit rates (continuing trend).
4. Forecasts 'No Initiation' Quit rates (counterfactual).
}
