% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_relapse.R
\name{estimate_relapse}
\alias{estimate_relapse}
\title{Estimate and Forecast Smoking Relapse}
\usage{
estimate_relapse(config, survey_data)
}
\description{
1. Prepares base relapse rates (Hawkins).
2. Forecasts the Age/Sex/IMD specific trend using `quit_forecast`.
3. Scales the Time-Since-Quit (TSQ) data using `relapse_forecast`.
4. Imputes data for ages < 18.
}
