% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ever_smoke.R
\name{ever_smoke}
\alias{ever_smoke}
\title{Summarise and project trends in ever-smoking  \lifecycle{maturing}}
\usage{
ever_smoke(data, time_horizon = 2100, num_bins = 7, model = "England")
}
\arguments{
\item{data}{Data table of individual characteristics.}

\item{time_horizon}{Integer - the last year for which to generate
predictions for ever-smoking at the index age.}

\item{num_bins}{Integer - the number of bins to create}

\item{model}{Character - whether running the code using HSE or SHeS data.
Fits a different model to trends depending
on dataset using.}
}
\value{
Returns two data tables:
\itemize{
\item "data_points", the proportions of ever-smokers at age 25 observed in the survey data;
\item "predicted_values", the model predictions up to the time horizon.
}
}
\description{
Calculates the expected proportion of ever-smokers, and
fits a linear trend over time to the proportion of people who have ever smoked in the
age category 25-34 years.
}
\details{
Uses the 'survey' package to fit a model to the trends in
ever-smoking that accounts for survey sampling
error and survey design.

The linear model is fitted to a binary response variable, 1 = ever smoker,
0 = never smoker and estimates
ariation in the trend
by sex and quintiles of the Index of Multiple Deprivation.
The fitted model is then used to predict /
extrapolate the trend up to the year specified by 'time horizon'.
}
\examples{

\dontrun{

test_data_trend <- ever_smoke(data = hse_data)

}

}
