% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkins_relapse.R
\docType{data}
\name{hawkins_relapse}
\alias{hawkins_relapse}
\title{Long-term probabilities of smoking relapse}
\format{
A data table.
}
\source{
\href{https://academic.oup.com/ntr/article/12/12/1228/1021710}{James Hawkins, William Hollingworth, Rona Campbell, Long-Term Smoking Relapse: A Study Using the British Household Panel Survey, Nicotine & Tobacco Research, Volume 12, Issue 12, December 2010, Pages 1228–1235, https://doi.org/10.1093/ntr/ntq175}
}
\usage{
hawkins_relapse
}
\description{
Probabilities of smoking relapse estimated from Hawkins 2010, who analysed the British Household Panel Survey.
}
\keyword{datasets}
