% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_adj.R
\name{init_adj}
\alias{init_adj}
\title{Adjust probabilities of ever-smoking}
\usage{
init_adj(
  init_data,
  ever_smoke_data,
  ref_age = 30,
  cohorts = 1971:2020,
  period_start = 2001,
  period_end = 2018
)
}
\arguments{
\item{init_data}{Data table - raw estimates of the probabilities of 
ever-smoking by age, sex and IMD quintile.}

\item{ever_smoke_data}{Data table - reference values for the proportion of ever-smokers.}

\item{ref_age}{Integer - the index age for calibration}

\item{cohorts}{Integer vector - the cohorts to be adjusted}

\item{period_start}{Integer - the first year of data}

\item{period_end}{Integer - the last year of data}
}
\value{
Returns an updated version of init_data with a new 
variable containing the adjusted values for
the probabilities of ever smoking. The data has also 
been filtered to only include cohorts for which it is possible
to make an adjustment.
}
\description{
Takes the probabilities of ever-smoking estimated from respondent recall and adjusts them
according to the reported proportions of ever-smokers at a reference age.
}
\details{
This is based on the method applied by \insertCite{holford2014patterns;textual}{smktrans}.
}
\examples{

\dontrun{

# Estimate the raw probabilities of ever-smoking
init_data_raw <- init_est(
  data = hse_data,
  strat_vars = c("sex", "imd_quintile")
)

# Estimate the reference values - observed proportions of ever-smokers
ever_smoke_data <- ever_smoke(
  data = hse_data,
  time_horizon = 2100
)

# Adjust the probabilities of ever-smoking
init_data_adj <- init_adj(
  init_data = copy(init_data_raw),
  ever_smoke_data = copy(ever_smoke_data$predicted_values),
  ref_age = 30,
  cohorts = 1971:2020, # 50 cohorts
  period_start = 2001,
  period_end = 2016
)


}

}
\references{
\insertRef{holford2014patterns}{smktrans}
}
