% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_est.R
\name{init_est}
\alias{init_est}
\title{Cohort specific smoking initiation}
\usage{
init_est(data, strat_vars = c("sex", "imd_quintile"))
}
\arguments{
\item{data}{Data table of individual characteristics. Must contain 'start_age'.}

\item{strat_vars}{Character vector of stratification variables.}
}
\value{
A summarized data.table of initiation probabilities by age/year/cohort.
}
\description{
Reconstructs longitudinal smoking histories from cross-sectional recall data.
}
