% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_dense.R
\name{p_dense}
\alias{p_dense}
\title{Convert probabilities of ever-smoking to age-specific probabilities}
\usage{
p_dense(
  data,
  cum_func_var,
  strat_vars = c("cohort", "sex", "imd_quintile"),
  lowest_year = 2003,
  max_year = 2100
)
}
\arguments{
\item{data}{Data table with cumulative probabilities.}

\item{cum_func_var}{Character - name of cumulative variable.}

\item{strat_vars}{Character vector - stratification variables.}

\item{lowest_year}{integer - start year filter.}

\item{max_year}{integer - end year filter.}
}
\description{
Converts the Cumulative Distribution Function (CDF) of ever-smoking 
into the Probability Density Function (PDF), which represents the 
probability of initiating smoking at a specific age.
}
