% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_relapse.R
\name{prep_relapse}
\alias{prep_relapse}
\title{Prepare long-term relapse probabilities}
\usage{
prep_relapse(
  data,
  hawkins_relapse = smktrans::hawkins_relapse,
  lowest_year = 2003,
  highest_year = 2018,
  youngest_age = 18
)
}
\arguments{
\item{data}{Data table of individual survey data.}

\item{hawkins_relapse}{Data table of Hawkins et al. odds ratios/probs.}

\item{lowest_year}{Integer.}

\item{highest_year}{Integer.}

\item{youngest_age}{Integer.}
}
\description{
Combines published estimates of long-term relapse (Hawkins 2010) with 
population survey data. It maps these probabilities onto the survey population's 
demographic structure (education, mental health, etc.) to derive 
aggregate relapse rates by Age, Sex, IMD, and Time Since Quit.
}
