% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_surv.R
\name{prep_surv}
\alias{prep_surv}
\title{Estimate socioeconomic differences in cohort survivorship}
\usage{
prep_surv(
  mx_data_hmd = smktrans::hmd_data_eng,
  mx_data_ons,
  min_age = 11,
  max_age = 89,
  min_year = 2003,
  max_year = 2018
)
}
\arguments{
\item{mx_data_hmd}{Data table of HMD mortality rates (1922+).}

\item{mx_data_ons}{Data table of ONS stratified mortality rates.}

\item{min_age, max_age, min_year, max_year}{Integers defining the scope.}
}
\description{
Combines long-term historical mortality rates (HMD) with recent socioeconomically 
stratified mortality data (ONS) to produce cohort survivorship curves (lx) 
stratified by Age, Sex, and IMD Quintile.
}
