% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_country_wrapper.R
\name{process_country}
\alias{process_country}
\title{Master Process Wrapper}
\usage{
process_country(config)
}
\arguments{
\item{config}{A named list containing all country-specific parameters}
}
\description{
Master Process Wrapper
}
