% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quit_est.R
\name{quit_est}
\alias{quit_est}
\title{Estimate smoking quit probabilities (The Flow Equation)}
\usage{
quit_est(
  trend_data,
  survivorship_data,
  mortality_data,
  relapse_data,
  initiation_data,
  min_age = 11,
  max_age = 89,
  min_year = 2003,
  max_year = 2018
)
}
\arguments{
\item{trend_data}{Output of trend_fit().}

\item{survivorship_data}{Output of prep_surv().}

\item{mortality_data}{Output of smoke_surv().}

\item{relapse_data}{Output of prep_relapse() (aggregated version).}

\item{initiation_data}{Output of init_adj().}

\item{min_age, max_age, min_year, max_year}{Integers.}
}
\description{
Calculates the probability of quitting required to balance the population 
stocks (Current, Former, Never) given the flows (Initiation, Relapse, Mortality).
}
\details{
Solves the equation:
N(t+1) = N(t) * Survival * (1 - Quit) + (Former * Relapse) + (Never * Init)
Rearranged to solve for Quit.
}
