% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quit_forecast.R
\name{quit_forecast}
\alias{quit_forecast}
\title{Forecast probabilities of smoking initiation, quitting and relapse}
\usage{
quit_forecast(
  data,
  forecast_var,
  forecast_type = c("continuing", "stationary"),
  cont_limit = NULL,
  oldest_year = 2003,
  youngest_age = 11,
  oldest_age = 88,
  age_cont_limit = 88,
  first_year = 2010,
  jump_off_year = 2015,
  time_horizon = 2050,
  smooth_rate_dim = c(3, 3),
  k_smooth_age = 3
)
}
\arguments{
\item{data}{Data table with input probabilities.}

\item{forecast_var}{Character - variable to forecast.}

\item{forecast_type}{"continuing" (linear trend) or "stationary" (constant).}

\item{cont_limit}{Integer - year where forecast becomes stationary.}

\item{oldest_year}{Integer - start of historical data.}

\item{youngest_age}{Integer - min age.}

\item{oldest_age}{Integer - max age.}

\item{first_year}{Integer - start year for trend fitting.}

\item{jump_off_year}{Integer - end year of historical data.}

\item{time_horizon}{Integer - end year of forecast.}

\item{smooth_rate_dim}{Vector - dimensions for raster smoothing (c(3,3)).}

\item{k_smooth_age}{Integer - knots for smoothing age component.}
}
\description{
Forecasts trends in transition probabilities using a Lee-Carter style 
Singular Value Decomposition (SVD) model.
}
\details{
The model assumes the logit of the probability can be decomposed into:
Logit(P_xt) = Alpha_x + Beta_x * Kappa_t
Where:
- Alpha_x: Average age profile
- Kappa_t: Time trend index
- Beta_x: Sensitivity of each age to the time trend
}
