% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relapse_forecast.R
\name{relapse_forecast}
\alias{relapse_forecast}
\title{Forecast relapse probabilities (Time-Since-Quit Stratified)}
\usage{
relapse_forecast(
  relapse_forecast_data,
  relapse_by_age_imd_timesincequit,
  jump_off_year = 2018
)
}
\arguments{
\item{relapse_forecast_data}{Data table - output from quit_forecast() (The trend).}

\item{relapse_by_age_imd_timesincequit}{Data table - detailed base rates from prep_relapse().}

\item{jump_off_year}{Integer.}
}
\description{
Applies the trends forecast by `quit_forecast` (which are only Age/Sex/IMD)
to the detailed Time-Since-Quit data. It calculates a scaling factor
from the "jump-off" year and applies it to future years.
}
