% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smktrans-package.R
\docType{package}
\name{smktrans-package}
\alias{smktrans}
\alias{smktrans-package}
\title{smktrans: smktrans: An R Package for estimating smoking state transition probabilities from cross-sectional survey data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimates the probabilities that individuals will initiate, quit or relapse to regular cigarette smoking from cross-sectional survey data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stapm-platform/smktrans}
  \item \url{https://stapm-platform.github.io/smktrans}
  \item Report bugs at \url{https://github.com/stapm-platform/smktrans/issues}
}

}
\author{
\strong{Maintainer}: Duncan Gillespie \email{duncan.gillespie@sheffield.ac.uk} (\href{https://orcid.org/0000-0003-3450-5747}{ORCID})

Authors:
\itemize{
  \item Alan Brennan \email{a.brennan@sheffield.ac.uk} (\href{https://orcid.org/0000-0002-1025-312X}{ORCID})
}

Other contributors:
\itemize{
  \item Laura Webster (\href{https://orcid.org/0000-0002-4307-9631}{ORCID}) (Original code and methodology) [contributor]
  \item Grace Leeming \email{g.leeming@sheffield.ac.uk} (\href{https://orcid.org/0000-0003-4836-7840}{ORCID}) (Code extension to Scotland) [contributor]
  \item Robin Purshouse (Methodological review and feedback) [contributor]
  \item Hazel Squires (Methodological review and feedback) [contributor]
  \item Shangshang Gu (Methodological review and feedback) [contributor]
}

}
\keyword{internal}
