% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smktrans-package.R
\docType{package}
\name{smktrans-package}
\alias{smktrans}
\alias{smktrans-package}
\title{smktrans: Smoking Initiation, Quit and Relapse Probabilities from Cross-sectional Survey Data}
\description{
Estimates the probabilities that individuals will initiate or quit regular cigarette smoking from cross-sectional survey data.
}
\author{
\strong{Maintainer}: Duncan Gillespie \email{duncan.gillespie@sheffield.ac.uk} (\href{https://orcid.org/0000-0003-3450-5747}{ORCID})

Authors:
\itemize{
  \item Laura Webster
  \item Alan Brennan \email{a.brennan@sheffield.ac.uk}
}

}
\keyword{internal}
