% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoke_surv.R
\name{smoke_surv}
\alias{smoke_surv}
\title{Estimate age-specific probabilities of death by smoking status}
\usage{
smoke_surv(
  data,
  diseases = tobalcepi::tob_disease_names,
  mx_data,
  min_age = 11,
  max_age = 89,
  min_year = 2003,
  max_year = 2018
)
}
\arguments{
\item{data}{Data table of individual characteristics (survey data).}

\item{diseases}{Character vector of disease names.}

\item{mx_data}{Data table of cause-specific mortality rates.}

\item{min_age, max_age, min_year, max_year}{Integers.}
}
\description{
Calculates survival probabilities (px) stratified by Smoking Status.
It uses the relative risks (RR) of smoking-related diseases to adjust
background mortality rates.
}
