% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_fit.R
\name{trend_fit}
\alias{trend_fit}
\title{Statistically model trends in smoking status}
\usage{
trend_fit(
  data,
  max_iterations = 1000,
  age_var = "age",
  year_var = "year",
  sex_var = "sex",
  smoker_state_var = "smk.state",
  imd_var = "imd_quintile",
  weight_var = "wt_int"
)
}
\arguments{
\item{data}{Data table of survey data.}

\item{max_iterations}{Integer.}

\item{age_var, year_var, sex_var, smoker_state_var, imd_var, weight_var}{Column names.}
}
\description{
Fits a multinomial logistic regression "response surface" to estimate 
Current/Former/Never status. The model includes high-order polynomials 
for Age and Year, and interactions with Sex and IMD.
}
