% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_excel_report.R
\name{write_excel_report}
\alias{write_excel_report}
\title{Write Transition Probability Estimates to Excel}
\usage{
write_excel_report(
  config,
  init_res,
  quit_res,
  relapse_res,
  net_init_dt,
  quit_no_init
)
}
\arguments{
\item{config}{Configuration list.}

\item{init_res}{Initiation results list.}

\item{quit_res}{Quitting results list.}

\item{relapse_res}{Relapse results list.}

\item{net_init_dt}{Net initiation data table.}
}
\description{
Generates a multi-sheet Excel workbook with formatted data and a cover sheet.
}
